Remove negative testing
Removed negative testing as scope didn't specify negative tests and there issues with the system when it came to logging in with incorrect details where the system would still successfully log in. Negative tests will be revisited at a later time once the main scope of the end to end tests are completed.

Signed-off-by: Mohammad Azwad Rashid <m.azwad.rashid@gmail.com>